#### PR description:

Tuning the looping logic so that we can remove the redundant looping of elements for link tests in PFBlockAlgo. This simple change should make the PFBlockAlgo [one of the most timing consuming part of PF] faster by a factor up to 1.8 with no change in outputs.

Using HLT's OnLine_HLT_GRun.py on:
/eos/cms/store/data/Run2018D/EphemeralHLTPhysics5/RAW/v1/000/324/897/00000/F5F13777-D664-C747-88AA-0459C1E35254.root
(1000 events)
**FastReport**
```
[Before]
CPU time avg.      when run  Real time avg.      when run     Alloc. avg.      when run   Dealloc. avg.      when run  Modules
        44.0 ms        44.0 ms        44.7 ms        44.7 ms       +3349 kB       +3349 kB       -2607 kB       -2607 kB    hltParticleFlowBlock
         0.9 ms        44.0 ms         0.9 ms        44.4 ms         +72 kB       +3434 kB         -55 kB       -2663 kB    hltParticleFlowBlockForTaus
         1.5 ms        43.0 ms         1.6 ms        45.2 ms        +112 kB       +3121 kB         -85 kB       -2381 kB    hltParticleFlowBlockReg

[With this PR]
        24.7 ms        24.7 ms        26.3 ms        26.3 ms       +3349 kB       +3349 kB       -2607 kB       -2607 kB    hltParticleFlowBlock
         0.5 ms        25.2 ms         0.6 ms        28.7 ms         +72 kB       +3438 kB         -56 kB       -2668 kB    hltParticleFlowBlockForTaus
         0.9 ms        24.5 ms         1.0 ms        27.8 ms        +112 kB       +3120 kB         -85 kB       -2382 kB    hltParticleFlowBlockReg
```

And, just for reference, if we also drop PS from PFBlockProducer for HLT as discussed in the TSG meeting, March 4th, (https://github.com/hatakeyamak/cmssw/pull/new/PFBlockProducerDropPS2)
```
        14.7 ms        14.7 ms        15.9 ms        15.9 ms       +1742 kB       +1742 kB       -1400 kB       -1400 kB    hltParticleFlowBlock
         0.3 ms        16.2 ms         0.4 ms        17.3 ms         +36 kB       +1752 kB         -29 kB       -1399 kB    hltParticleFlowBlockForTaus
         0.5 ms        12.6 ms         0.5 ms        12.9 ms         +52 kB       +1460 kB         -41 kB       -1146 kB    hltParticleFlowBlockReg
```

#### PR validation:

Check with 100 events from 2018 data that this change yields the identical list of PF candidates.

#### if this PR is a backport please specify the original PR and why you need to backport that PR:

This is not backport.

@bendavid @jsalfeld @jpata