@slava77 2020-02-27
code-check

if the bot lost the last attempt