@hatakeyamak 2020-03-02
I expected absolutely no difference, but I see some failures.

Most of reported differences are related to:
recoCandidateedmPtrsrecoJetTagInforecoIPTagInfos_pfImpactParameterTagInfos__(re)RECO_obj_m_data_ghostTrackWeight
recoCandidateedmPtrsrecoJetTagInforecoIPTagInfos_pfImpactParameterTagInfos__RECO_obj_m_data_distanceToGhostTrack_value
and some muon-related DQM plots, but I am not sure if I understand the reported differences. Any suggestions will be appreciated to move this forward.