@jpata 2020-03-02
Hi @hatakeyamak, just a quick question - did the number of PFBlocks stay the same before and after the change? I wonder if we would benefit from automatic tests of the code on one event with full verbose printout (LogDebug) - this can easily be diffed. I could set this up.