@hatakeyamak 2020-03-02
> Hi @hatakeyamak, just a quick question - did the number of PFBlocks stay the same before and after the change? I wonder if we would benefit from automatic tests of the code on one event with full verbose printout (LogDebug) - this can easily be diffed. I could set this up.

Thanks. Yes, the number of PFBlocks stays the same in the test I looked at. The list of PF candidates also stays the same as far as I can tell, but the ordering of candidates in the collection change, which I think is due to the order of PFBlocks change, and this is probably expected because the order of union-find algo in PFBlockAlgo is not kept the same with this update. Check of this assumption could be very useful as you say.

All these changes shouldn't be a problem as long as physical observables are at the end sorted in some ways. I just cannot quite connect this expectation with reported failures, as I don't fully understand all the reported failed tests.