@cmsbuild 2020-03-06
+1
Tested at: 13ff4d8fb9835e101fce8cff1ec451ca4dc09519
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-39a3bd/5040/summary.html
CMSSW: CMSSW_11_1_X_2020-03-05-2300
SCRAM_ARCH: slc7_amd64_gcc820