@hatakeyamak 2020-03-09
In my investigation of
```
recoCandidateedmPtrsrecoJetTagInforecoIPTagInfos_pfImpactParameterTagInfos__(re)RECO_obj_m_data_ghostTrackWeight
recoCandidateedmPtrsrecoJetTagInforecoIPTagInfos_pfImpactParameterTagInfos__RECO_obj_m_data_distanceToGhostTrack_value
```
I see some small changes in:
```
vector<reco::Track>                   "pfImpactParameterTagInfos"   "ghostTracks"     "(re)RECO"    
vector<reco::IPTagInfo<vector<edm::Ptr<reco::Candidate> >,reco::JetTagInfo> >    "pfImpactParameterTagInfos"   ""                "(re)RECO"    
```
For example, in output of 136.874_RunEGamma2018C, 100 events, I see 3 cases of tiny differences in ghostTracks:
```
108066c108066
< tracks: pt   0.6561851033277 eta  -0.98355321390117 phi   1.87568111755637 
---
> tracks: pt   0.6540341172427 eta  -0.98789902418708 phi   1.87540677496420 

118614c118614
< tracks: pt   0.8130750166088 eta  -0.66572216238707 phi  -2.40236842625495 
---
> tracks: pt   0.8130774022123 eta  -0.66571723565971 phi  -2.40238250896182 

208658c208658
< tracks: pt   0.6187283287962 eta   1.05984628567672 phi  -2.38975964612986 
---
> tracks: pt   0.6167970203661 eta   1.06382187233677 phi  -2.39066671374844 
```
My impression is that ghosttrack reco is very sensitive to numerical differences, which can be caused by changes of pf candidates order, but the differences is very small. I doubt it's something we need to worry about.

Regarding some changes in DQM plots, I think they are showing up for plots involving
vector<reco::PFCandidate>             "pfIsolatedMuonsEI"         ""                "(re)RECO"
and it's because these muons are not sorted in any particular way [and there is no better or worse in how this collection was defined before or after this PR]. I think the better way can be explored after carefully check also how other pf*EI collections are also defined/sorted, and it can be postponed to another PR.